requirements.txt
openai==0.28.1
chromadb
python-dotenv==1.0.0
tiktoken==0.5.1
langchain==0.1.0
azure-functions
azure-ai-formrecognizer==3.2.0
azure-storage-blob==1.3
azure-storage-queue==12.5.0
azure-search-documents==11.4.0
streamlit==1.29.0
numpy==1.26.1
pandas
thrift==0.16.0
openpyxl==3.1.2
pyhive
pypdf
pymongo